\name{justice}
\alias{justice}
\docType{data}
\title{Example data set on employees' justice perceptions and satisfaction}
\description{
This data set contains simulated data for employees nested within organizations, featuring employees' sex, ratings on individual justice orientation and ratings on job satisfaction.
In addition, the data set features scores for justice climate in each organization (defined at the level of organizations, level 2).
Different organizations are denoted by the variable \code{id}.

The data were simulated based on the results by Liao and Rupp (2005), as well as the secondary analyses of the same data given in Mathieu, Aguinis, Culpepper, and Chen, (2012).
}
\format{A data frame containing 1400 observations on 4 variables.}
\usage{data(justice)}
\references{
Liao, H., & Rupp, D. E. (2005). The impact of justice climate and justice orientation on work outcomes: A cross-level multifoci framework. \emph{Journal of Applied Psychology}, 90, 242.256.

Mathieu, J. E., Aguinis, H., Culpepper, S. A., & Chen, G. (2012). Understanding and estimating the power to detect cross-level interaction effects in multilevel modeling. \emph{Journal of Applied Psychology}, 97, 951-966.
}
\keyword{datasets}
