\name{leadership}
\alias{leadership}
\docType{data}
\title{Example data set on leadership style and job satisfaction}
\description{
Data set based on the data simulated by Paul D. Bliese as described in Klein et al. (2000) with slight modifications.
The data set consists of 750 employees, nested within 50 work groups, and includes employees' ratings on negative leadership style, job satisfaction, and workload as well as a measure for each work group's cohesion.

The original data set is available in the \code{multilevel} package and was modified by (a) transforming workload into a categorical variable, (b) transforming cohesion into a group-level variable, and (c) by inducing missing values.
}
\usage{data(leadership)}
\format{A data frame containing 750 observations on 5 variables.}
\references{

Bliese, P. D. (2013). multilevel: Multilevel functions (Version 2.5) [Computer software]. Retrieved from \code{http://CRAN.R-project.org/package=multilevel}

Klein, K. J., Bliese, P. D., Kozlowski, S. W. J., Dansereau, F., Gavin, M. B., Griffin, M. A., ... Bligh, M. C. (2000). Multilevel analytical techniques: Commonalities, differences, and continuing questions. In K. J. Klein & S. W. J. Kozlowski (Eds.), \emph{Multilevel theory, research, and methods in organizations: Foundations, extensions, and new directions} (pp. 512-553). San Francisco, CA: Jossey-Bass.

}
\keyword{datasets}
