\name{studentratings}
\alias{studentratings}
\docType{data}
\title{Example data set on student's ratings and achievement}
\description{
This is data set has been simulated to mimic typical data from educational research.
The data set features student's ratings of their teachers' behavior (i.e., disciplinary problems in mathematics and language class) and their general learning environment (school climate), as well as mathematics and language achievement scores, and scores for socio-economic state and cognitive ability.

In addition, the data set features the ID of 50 different schools (i.e., clusters), the biological sex of all students, and a broad, additional grouping factor. 
Different amounts of missing data have been inserted into the data set in a completely random fashion.

All data are completely fictional, and thus should not be interpreted as if they were real.
}
\usage{data(studentratings)}
\format{A data frame containing 750 observations on 10 variables.}
\keyword{datasets}
